🗃️ Add migrations and seeds
