+++ seeds/0000_empty_tables.js
@@ -1,3 +1,4 @@
 export async function seed(knex) {
+  await knex("user").del();
   await knex("story").del();
 }
